Re-organize fn_decls module
