+++ src/compiler/typescript_to_rust_redesign/azle_generate/src/azle_ast/mod.rs
@@ -1,0 +1,9 @@
+pub enum CanisterMethodType {
+    Heartbeat,
+    Init,
+    InspectMessage,
+    PostUpgrade,
+    PreUpgrade,
+    Query,
+    Update,
+}
