+++ src/compiler/typescript_to_rust_redesign/azle_generate/src/generators/canister_methods/system/heartbeat.rs
@@ -1,8 +1,6 @@
 use crate::{
-    generators::canister_methods::method_body::generate_call_to_js_function,
-    utils::fn_decls::{
-        get_canister_method_type_fn_decls, get_fn_decl_function_name, CanisterMethodType,
-    },
+    azle_ast::CanisterMethodType,
+    generators::canister_methods::method_body::generate_call_to_js_function, ts_ast,
 };
 use quote::{format_ident, quote};
 use swc_ecma_ast::Program;
@@ -10,8 +8,10 @@ use swc_ecma_ast::Program;
 pub fn generate_canister_method_system_heartbeat(
     programs: &Vec<Program>,
 ) -> proc_macro2::TokenStream {
-    let heartbeat_fn_decls =
-        get_canister_method_type_fn_decls(programs, &CanisterMethodType::Heartbeat);
+    let heartbeat_fn_decls = ts_ast::program::get_canister_method_type_fn_decls(
+        programs,
+        &CanisterMethodType::Heartbeat,
+    );
 
     if heartbeat_fn_decls.len() > 1 {
         panic!("Only one Heartbeat function can be defined");
@@ -20,7 +20,7 @@ pub fn generate_canister_method_system_heartbeat(
     let heartbeat_fn_decl_option = heartbeat_fn_decls.get(0);
 
     if let Some(heartbeat_fn_decl) = heartbeat_fn_decl_option {
-        let function_name = get_fn_decl_function_name(heartbeat_fn_decl);
+        let function_name = ts_ast::fn_decl::get_fn_decl_function_name(heartbeat_fn_decl);
         let rust_function_name_ident = format_ident!("_azle_heartbeat_{}", function_name);
 
         let call_to_heartbeat_js_function = generate_call_to_js_function(heartbeat_fn_decl);
