+++ src/compiler/typescript_to_rust_redesign/azle_generate/src/generators/canister_methods/system/init.rs
@@ -1,12 +1,13 @@
 use crate::generators::ic_object::generate_ic_object;
 use crate::{
+    azle_ast::CanisterMethodType,
     generators::canister_methods::{
         functions::{
             generate_param_name_idents, generate_param_types, generate_params_token_stream,
         },
         method_body::generate_call_to_js_function,
     },
-    utils::fn_decls::{get_canister_method_type_fn_decls, CanisterMethodType},
+    ts_ast,
 };
 use quote::quote;
 use swc_ecma_ast::{FnDecl, Program};
@@ -14,7 +15,8 @@ use swc_ecma_ast::{FnDecl, Program};
 pub fn generate_canister_method_system_init(programs: &Vec<Program>) -> proc_macro2::TokenStream {
     let ic_object = generate_ic_object(programs);
 
-    let init_fn_decls = get_canister_method_type_fn_decls(programs, &CanisterMethodType::Init);
+    let init_fn_decls =
+        ts_ast::program::get_canister_method_type_fn_decls(programs, &CanisterMethodType::Init);
 
     if init_fn_decls.len() > 1 {
         panic!("Only one Init function can be defined");
