+++ src/compiler/typescript_to_rust_redesign/azle_generate/src/generators/canister_methods/system/inspect_message.rs
@@ -1,8 +1,6 @@
 use crate::{
-    generators::canister_methods::method_body::generate_call_to_js_function,
-    utils::fn_decls::{
-        get_canister_method_type_fn_decls, get_fn_decl_function_name, CanisterMethodType,
-    },
+    azle_ast::CanisterMethodType,
+    generators::canister_methods::method_body::generate_call_to_js_function, ts_ast,
 };
 use quote::{format_ident, quote};
 use swc_ecma_ast::Program;
@@ -10,8 +8,10 @@ use swc_ecma_ast::Program;
 pub fn generate_canister_method_system_inspect_message(
     programs: &Vec<Program>,
 ) -> proc_macro2::TokenStream {
-    let inspect_message_fn_decls =
-        get_canister_method_type_fn_decls(programs, &CanisterMethodType::InspectMessage);
+    let inspect_message_fn_decls = ts_ast::program::get_canister_method_type_fn_decls(
+        programs,
+        &CanisterMethodType::InspectMessage,
+    );
 
     if inspect_message_fn_decls.len() > 1 {
         panic!("Only one InspectMessage function can be defined");
@@ -20,7 +20,7 @@ pub fn generate_canister_method_system_inspect_message(
     let inspect_message_fn_decl_option = inspect_message_fn_decls.get(0);
 
     if let Some(inspect_message_fn_decl) = inspect_message_fn_decl_option {
-        let function_name = get_fn_decl_function_name(inspect_message_fn_decl);
+        let function_name = ts_ast::fn_decl::get_fn_decl_function_name(inspect_message_fn_decl);
         let rust_function_name_ident = format_ident!("_azle_inspect_message_{}", function_name);
 
         let call_to_inspect_message_js_function =
