+++ src/compiler/typescript_to_rust_redesign/azle_generate/src/ts_ast/module.rs
@@ -1,0 +1,18 @@
+use swc_ecma_ast::{ExportDecl, Module, ModuleDecl};
+
+pub fn get_export_decls(module: &Module) -> Vec<ExportDecl> {
+    let module_decls: Vec<ModuleDecl> = module
+        .body
+        .iter()
+        .filter(|module_item| module_item.is_module_decl())
+        .map(|module_item| module_item.as_module_decl().unwrap().clone())
+        .collect();
+
+    let export_decls: Vec<ExportDecl> = module_decls
+        .iter()
+        .filter(|module_decl| module_decl.is_export_decl())
+        .map(|module_decl| module_decl.as_export_decl().unwrap().clone())
+        .collect();
+
+    export_decls
+}
